Add traveler Interaction
This commit adds the functionality for the traveler interaction to the project. After logging in the traveler now has the ability to create a trip request. After selecting the options for the trip the user will have an option to confirm. Once the trip is confirmed it is then added to the TRIPS database. In a future issue I plan on having the trips refresh after the new trip is added so that the trip is showing after it is submitted in the travelers trips section.