Update travis.yml file
At this time the TravisCI is succesfully linting once a pull request is opened. I have also added the DEPLOY instructions. The tests are still passing so far with me seeing the message:
"Skipping a deployment with the pages provider because the current build is a pull request."
My hope is that since everything appears to be ok once I merge the pull request travisCI will push this build to my gh-pages branch.