Update traveler dashboard
This commit updates the functionality on the traveler dashboard page. Now whenever a traveler clicks 'submit' and submits a trip request, the trip is added to the travelers Upcoming Trips array, which then renders it on the page after it is submitted. This way the traveler does not have to log out and back in / refresh in order to see their new trip request.