Update password input and style
This commit updates the project so that the password is shown as ****'s instead of being visible. After typing in the password the user is also able to press enter to login.
This commit also adds a small amount of margin to the logout button and padding to the heading on the trip request confirmation page. Also updating meta tag info to try and have page render correctly on LinkedIn.