Correct delete trip functionality for agent dashboard
This commit corrects the functionality for deleting a trip on the agent dashboard. On the traveler dashboard the traveler is able to submit a trip request and then if they want to submit another one they are able to input the info and see it to confirm, but having trouble with the POST request giving a 422 error when trying to submit JUST the 2nd trip request. Will continue to investigate.